cleaning up url rewrite
