+++ samples/kotlin/umaaas-quickstart/backend/src/main/resources/application.yaml
@@ -6,5 +6,4 @@ ktor:
         port: 8080
         watch:
             - classes
-            - resources
     development: true
